+++ plenum/test/waits.py
@@ -67,7 +67,9 @@ def expectedPoolInterconnectionTime(nodeCount):
     # bug (`'str' object has no attribute 'keys'`) which supposed to be
     # fixed in the 3pcbatch feature
     # https://evernym.atlassian.net/browse/SOV-995
-    return interconnectionCount * nodeConnectionTimeout + \
+    # multiply by 2 because we need to re-create connections which can be done on a second re-try only
+    # (we may send pings on some of the re-tries)
+    return 2 * interconnectionCount * nodeConnectionTimeout + \
            KITZStack.RETRY_TIMEOUT_RESTRICTED
